feat: working generate sdk command
chore: generate sdk clippy changes

feat: fix merge conflict

fix test
