+++ apps/framework-cli/src/framework/sdk.rs
@@ -1,0 +1,1 @@
+pub mod ingest;
