+++ apps/framework-cli/src/infrastructure/console.rs
@@ -14,9 +14,8 @@ use std::sync::Arc;
 
 use tokio::net::TcpStream;
 
-use crate::framework::controller::{
-    schema_file_path_to_ingest_route, FrameworkObjectVersions, SchemaVersion,
-};
+use crate::framework::controller::schema_file_path_to_ingest_route;
+use crate::framework::core::code_loader::{FrameworkObjectVersions, SchemaVersion};
 use crate::framework::data_model::schema::DataModel;
 use crate::infrastructure::olap;
 use crate::infrastructure::olap::clickhouse::model::ClickHouseSystemTable;
