+++ apps/framework-cli/src/infrastructure/olap/clickhouse/version_sync.rs
@@ -1,4 +1,4 @@
-use crate::framework::controller::{FrameworkObject, FrameworkObjectVersions};
+use crate::framework::core::code_loader::{FrameworkObject, FrameworkObjectVersions};
 use crate::framework::data_model::schema::DataModel;
 use crate::infrastructure::olap::clickhouse::model::{
     ClickHouseColumn, ClickHouseColumnType, ClickHouseTable,
